Implement a `WellKnownTypesSerializer`
When task input properties are serialized, we use a map serializer, for which the key serializer is optimized (uses a `String`)
but the value serializer is the generic, default, serializer, which will use Java serialization. That, because the map may
contain different object types. This commit introduces a new value serializer which first checks if the type we're trying
to serialize is "well known" (typically, `String`, `Long`, ...) and then it will use an optimized serializer instead.